# backend/src/app/generated/openapi.yaml
openapi: 3.0.3
info:
  description: |
    Public-first API for browsing latest public notes and managing a personal
    notebook of plain text notes (for anonymous and regular users). Authentication
    is via Firebase ID tokens passed as `Authorization: Bearer <token>`.

    ## Real-time Features
    This API includes WebSocket integration for real-time comment notifications.
    When comments are posted, they are automatically broadcasted to all connected
    WebSocket clients for live updates.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Simple Note Application API
  version: 1.0.0
servers:
- description: Production (placeholder)
  url: https://api.myapp.test
- description: Staging (placeholder)
  url: https://staging-api.myapp.test
...
paths:
...
  /notes/{id}/comments:
    get:
      operationId: getPublicNoteComments
      parameters:
      - description: Note ID (UUIDv4)
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/comments-list-response"
          description: List of comments for the note
        "404":
          $ref: "#/components/responses/not-found"
        "422":
          $ref: "#/components/responses/validation-error"
      summary: Get comments for a public note
      tags:
      - Comments
    post:
      description: |
        Creates a new comment on a public note. Requires authentication.
        Automatically broadcasts the new comment to all connected WebSocket clients.
      operationId: createPublicNoteComment
      parameters:
      - description: Note ID (UUIDv4)
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createPublicNoteComment_request"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/comment-response"
          description: Comment created successfully
        "400":
          $ref: "#/components/responses/validation-error"
        "401":
          $ref: "#/components/responses/unauthorized"
        "404":
          $ref: "#/components/responses/not-found"
        "422":
          $ref: "#/components/responses/validation-error"
      security:
      - BearerAuth: []
      summary: Create a comment on a public note
      tags:
      - Comments
...
  /me/notes/{id}/comments:
    get:
      operationId: getPrivateNoteComments
      parameters:
      - description: Note ID (UUIDv4)
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/comments-list-response"
          description: List of comments for the private note
        "401":
          $ref: "#/components/responses/unauthorized"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not-found"
        "422":
          $ref: "#/components/responses/validation-error"
      security:
      - BearerAuth: []
      summary: Get comments for my private note
      tags:
      - Comments
    post:
      description: |
        Creates a new comment on a private note owned by the authenticated user.
        Automatically broadcasts the new comment to all connected WebSocket clients.
      operationId: createPrivateNoteComment
      parameters:
      - description: Note ID (UUIDv4)
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createPublicNoteComment_request"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/comment-response"
          description: Comment created successfully
        "400":
          $ref: "#/components/responses/validation-error"
        "401":
          $ref: "#/components/responses/unauthorized"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not-found"
        "422":
          $ref: "#/components/responses/validation-error"
      security:
      - BearerAuth: []
      summary: Create a comment on my private note
      tags:
      - Comments
...
  /me:
    get:
      operationId: getUserProfile
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user-profile-response"
          description: User profile
        "401":
          $ref: "#/components/responses/unauthorized"
        "403":
          $ref: "#/components/responses/forbidden"
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - User Profile
    patch:
      operationId: updateUserProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/updateUserProfile_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user-profile-response"
          description: Updated profile
        "401":
          $ref: "#/components/responses/unauthorized"
        "403":
          $ref: "#/components/responses/forbidden"
        "422":
          $ref: "#/components/responses/validation-error"
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - User Profile
  /auth/anonymous-login:
    post:
      operationId: authenticateAnonymous
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/auth-result-response"
          description: Anonymous session validated; user ensured in DB
        "401":
          $ref: "#/components/responses/unauthorized"
        "403":
          $ref: "#/components/responses/forbidden"
        "422":
          $ref: "#/components/responses/validation-error"
      security:
      - BearerAuth: []
      summary: Authenticate anonymous user (validate Firebase anonymous token)
      tags:
      - Authentication
  /auth/login:
    get:
      operationId: loginRegularUser
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/auth-result-response"
          description: Regular user session validated (and created if first time)
        "401":
          $ref: "#/components/responses/unauthorized"
        "403":
          $ref: "#/components/responses/forbidden"
      security:
      - BearerAuth: []
      summary: Authenticate regular user (validate Firebase token; may insert user)
      tags:
      - Authentication
  /auth/anonymous-promote:
    post:
      operationId: promoteAnonymousUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/anonymous-promote-request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/auth-result-response"
          description: Anonymous account promoted to regular user
        "401":
          $ref: "#/components/responses/unauthorized"
        "403":
          $ref: "#/components/responses/forbidden"
        "422":
          $ref: "#/components/responses/validation-error"
      security:
      - BearerAuth: []
      summary: Promote anonymous user to regular account
      tags:
      - Authentication
components:
...
  responses:
    unauthorized:
      content:
        application/json:
          examples:
            example:
              value:
                status: error
                error:
                  code: UNAUTHORIZED
                  message: Missing or invalid token
          schema:
            $ref: "#/components/schemas/error-response"
      description: Unauthorized
    forbidden:
      content:
        application/json:
          examples:
            example:
              value:
                status: error
                error:
                  code: FORBIDDEN
                  message: Insufficient permissions
          schema:
            $ref: "#/components/schemas/error-response"
      description: Forbidden
    validation-error:
      content:
        application/json:
          examples:
            example:
              value:
                status: error
                error:
                  code: VALIDATION_ERROR
                  message: Invalid request payload
                  field: content
          schema:
            $ref: "#/components/schemas/error-response"
      description: Validation error
    not-found:
      content:
        application/json:
          examples:
            example:
              value:
                status: error
                error:
                  code: NOT_FOUND
                  message: Resource not found
          schema:
            $ref: "#/components/schemas/error-response"
      description: Not Found
  schemas:
...
    error-response:
      additionalProperties: false
      description: Standard error wrapper
      example:
        error:
          code: code
          field: field
          message: message
        status: error
      properties:
        status:
          enum:
          - error
          title: status
          type: string
        error:
          $ref: "#/components/schemas/error_response_error"
      required:
      - error
      - status
      title: error-response
      type: object
...
    comment:
      additionalProperties: false
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        postId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        content: content
        username: username
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          description: Unique comment identifier
          format: uuid
          title: id
          type: string
        content:
          description: Comment text content
          maxLength: 1000
          minLength: 1
          title: content
          type: string
        username:
          description: Display name or email of the comment author
          title: username
          type: string
        createdAt:
          description: Comment creation timestamp
          format: date-time
          title: createdAt
          type: string
        updatedAt:
          description: Comment last update timestamp
          format: date-time
          title: updatedAt
          type: string
        postId:
          description: ID of the note this comment belongs to
          format: uuid
          title: postId
          type: string
      required:
      - content
      - createdAt
      - id
      - postId
      - updatedAt
      - username
      title: comment
      type: object
    comments-list-response:
      additionalProperties: false
      example:
        data:
          comments:
          - createdAt: 2000-01-23T04:56:07.000+00:00
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            postId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            content: content
            username: username
            updatedAt: 2000-01-23T04:56:07.000+00:00
          - createdAt: 2000-01-23T04:56:07.000+00:00
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            postId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            content: content
            username: username
            updatedAt: 2000-01-23T04:56:07.000+00:00
          count: 5
          postId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        success: true
        message: Comments retrieved successfully
      properties:
        success:
          example: true
          title: success
          type: boolean
        data:
          $ref: "#/components/schemas/comments_list_response_data"
        message:
          example: Comments retrieved successfully
          title: message
          type: string
      required:
      - data
      - success
      title: comments-list-response
      type: object
    comment-response:
      additionalProperties: false
      example:
        data:
          createdAt: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          postId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          content: content
          username: username
          updatedAt: 2000-01-23T04:56:07.000+00:00
        success: true
        message: Comment created successfully
      properties:
        success:
          example: true
          title: success
          type: boolean
        data:
          $ref: "#/components/schemas/comment"
        message:
          example: Comment created successfully
          title: message
          type: string
      required:
      - data
      - success
      title: comment-response
      type: object
...
    user-profile:
      additionalProperties: false
      description: Basic user profile for regular users
      example:
        uid: uid
        createdAt: 2000-01-23T04:56:07.000+00:00
        isAnonymous: true
        avatarUrl: https://openapi-generator.tech
        displayName: displayName
        email: email
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        uid:
          description: Firebase UID
          title: uid
          type: string
        displayName:
          title: displayName
          type: string
        email:
          format: email
          nullable: true
          title: email
          type: string
        avatarUrl:
          format: uri
          nullable: true
          title: avatarUrl
          type: string
        isAnonymous:
          title: isAnonymous
          type: boolean
        createdAt:
          format: date-time
          title: createdAt
          type: string
        updatedAt:
          format: date-time
          title: updatedAt
          type: string
      required:
      - createdAt
      - displayName
      - isAnonymous
      - uid
      - updatedAt
      title: user-profile
      type: object
    user-profile-response:
      additionalProperties: false
      example:
        data:
          uid: uid
          createdAt: 2000-01-23T04:56:07.000+00:00
          isAnonymous: true
          avatarUrl: https://openapi-generator.tech
          displayName: displayName
          email: email
          updatedAt: 2000-01-23T04:56:07.000+00:00
        status: success
      properties:
        status:
          enum:
          - success
          title: status
          type: string
        data:
          $ref: "#/components/schemas/user-profile"
      required:
      - data
      - status
      title: user-profile-response
      type: object
    auth-result-response:
      additionalProperties: false
      example:
        data:
          isAnonymous: true
          created: true
          user:
            uid: uid
            createdAt: 2000-01-23T04:56:07.000+00:00
            isAnonymous: true
            avatarUrl: https://openapi-generator.tech
            displayName: displayName
            email: email
            updatedAt: 2000-01-23T04:56:07.000+00:00
          wasPromoted: true
        status: success
      properties:
        status:
          enum:
          - success
          title: status
          type: string
        data:
          $ref: "#/components/schemas/auth_result_response_data"
      required:
      - data
      - status
      title: auth-result-response
      type: object
    anonymous-promote-request:
      additionalProperties: false
      properties:
        anonymous_firebase_uuid:
          description: The Firebase UID of the anonymous user to be promoted
          format: uuid
          title: anonymous_firebase_uuid
          type: string
      required:
      - anonymous_firebase_uuid
      title: anonymous-promote-request
      type: object
...
    createPublicNoteComment_request:
      additionalProperties: false
      properties:
        content:
          description: Comment text content
          maxLength: 1000
          minLength: 1
          title: content
          type: string
      required:
      - content
      title: createPublicNoteComment_request
      type: object
...
    updateUserProfile_request:
      additionalProperties: false
      properties:
        displayName:
          maxLength: 60
          title: displayName
          type: string
        email:
          format: email
          title: email
          type: string
        avatarUrl:
          format: uri
          title: avatarUrl
          type: string
      title: updateUserProfile_request
      type: object
...
    error_response_error:
      example:
        code: code
        field: field
        message: message
      properties:
        code:
          title: code
          type: string
        message:
          title: message
          type: string
        field:
          nullable: true
          title: field
          type: string
      required:
      - code
      - message
      title: error_response_error
      type: object
    comments_list_response_data:
      example:
        comments:
        - createdAt: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          postId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          content: content
          username: username
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - createdAt: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          postId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          content: content
          username: username
          updatedAt: 2000-01-23T04:56:07.000+00:00
        count: 5
        postId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        comments:
          description: List of comments for the note
          items:
            $ref: "#/components/schemas/comment"
          title: comments
          type: array
        count:
          description: Total number of comments
          example: 5
          title: count
          type: integer
        postId:
          description: ID of the note these comments belong to
          format: uuid
          title: postId
          type: string
      required:
      - comments
      - count
      - postId
      title: comments_list_response_data
      type: object
...
    auth_result_response_data:
      example:
        isAnonymous: true
        created: true
        user:
          uid: uid
          createdAt: 2000-01-23T04:56:07.000+00:00
          isAnonymous: true
          avatarUrl: https://openapi-generator.tech
          displayName: displayName
          email: email
          updatedAt: 2000-01-23T04:56:07.000+00:00
        wasPromoted: true
      properties:
        user:
          $ref: "#/components/schemas/user-profile"
        isAnonymous:
          title: isAnonymous
          type: boolean
        created:
          description: Indicates if a new DB user record was created during auth
          title: created
          type: boolean
        wasPromoted:
          description: Indicates if an anonymous account was promoted
          title: wasPromoted
          type: boolean
      required:
      - user
      title: auth_result_response_data
      type: object
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      description: "Firebase ID token in `Authorization: Bearer <token>` header"
      scheme: bearer
      type: http